Adding settings and splitting files
This update splits the app into multiple files per best practice so it's easier to maintain.

It also adds a nav panel with a settings window to input your api key. The nav panel also has dummy elements and as I am preparing to support multiple conversations in the future.